[Impeller] Use 32 bit Gaussian function in the 2-pass blur (flutter/engine#42069)
Resolves https://github.com/flutter/flutter/issues/126487.

Increases the 2-pass blur quality and distribution limit.

It turns out sigma was breaking down beyond ~255 (moderately high, but
not unreasonable for users to rely on). The Gaussian function computes
sigma^2, and half precision floats only have 5 bit exponents and
overflow for numbers above 65k.

Coincidentally, this also returns us to a state where we look a lot more
like Skia's blurs for larger blur sigmas. Medium blurs have much less
visual banding (although it's still there if you look closely). I
suspect half precision isn't really enough for tracking the integral.

Unfortunately, this means our SIMD pipelining isn't going to be as good.
I'll be interested in watching the blur-driven benchmarks for the perf
hit.

+++ engine/src/flutter/impeller/entity/shaders/gaussian_blur/gaussian_blur.glsl
@@ -27,7 +27,7 @@ uniform BlurInfo {
   // host-side, but both are useful controls here. Sigma (pixels per standard
   // deviation) is used to define the gaussian function itself, whereas the
   // radius is used to limit how much of the function is integrated.
-  float16_t blur_sigma;
+  float blur_sigma;
   float16_t blur_radius;
 }
 blur_info;
@@ -62,7 +62,11 @@ void main() {
 
   for (float16_t i = -blur_info.blur_radius; i <= blur_info.blur_radius;
        i += 2.0hf) {
-    float16_t gaussian = IPGaussian(i, blur_info.blur_sigma);
+    // Use the 32 bit Gaussian function because the 16 bit variation results in
+    // quality loss/visible banding. Also, 16 bit variation internally breaks
+    // down at a moderately high (but still reasonable) blur sigma of >255 when
+    // computing sigma^2 due to the exponent only having 5 bits.
+    float16_t gaussian = float16_t(IPGaussian(float(i), blur_info.blur_sigma));
     gaussian_integral += gaussian;
     total_color +=
         gaussian * Sample(texture_sampler,  // sampler
